+++ CHANGES.rst
@@ -1,6 +1,12 @@
 0.5.0 (2021-12-08)
 ==================
 
+general
+-------
+
+- Added regression tests for SOC-604. [#381]
+
+
 linearity
 ---------